Add files via upload
I wrote some functions to automate collecting the data.
They are not perfect but they get the data down and into .csv files.
How to run it:

Run:

get_data()

It will ask you how many crpytos you want to retrieve data for ... I have tested for 10 and it works fine.

It next asks you for the coin id's for the ten crypto coins.  I look up the API id on Coin Gecko and use that. Be sure to separate them by a space, if you use a comma it will choke.

It next asks for the beginning data. Use format: yyyy-mm-dd

It next asks for the end date of the data, use the format as above.

hit enter and wait .... although it doesn't take very loonggg

the .csv files will appear in the documents list at the left side of jupyter notebooks

It prints "Task complete! when it is done.

The next thing that need to be done with this is to automate concating into tables.